io: remove adapters from main crate.
