+++ embedded-io/src/lib.rs
@@ -8,9 +8,6 @@ use core::fmt;
 #[cfg(feature = "alloc")]
 extern crate alloc;
 
-#[cfg(feature = "std")]
-pub mod adapters;
-
 mod impls;
 
 /// Enumeration of possible methods to seek within an I/O object.
@@ -24,6 +21,30 @@ pub enum SeekFrom {
     End(i64),
     /// Sets the offset to the current position plus the specified number of bytes.
     Current(i64),
+}
+
+#[cfg(feature = "std")]
+#[cfg_attr(docsrs, doc(cfg(feature = "std")))]
+impl From<SeekFrom> for std::io::SeekFrom {
+    fn from(pos: SeekFrom) -> Self {
+        match pos {
+            SeekFrom::Start(n) => std::io::SeekFrom::Start(n),
+            SeekFrom::End(n) => std::io::SeekFrom::End(n),
+            SeekFrom::Current(n) => std::io::SeekFrom::Current(n),
+        }
+    }
+}
+
+#[cfg(feature = "std")]
+#[cfg_attr(docsrs, doc(cfg(feature = "std")))]
+impl From<std::io::SeekFrom> for SeekFrom {
+    fn from(pos: std::io::SeekFrom) -> SeekFrom {
+        match pos {
+            std::io::SeekFrom::Start(n) => SeekFrom::Start(n),
+            std::io::SeekFrom::End(n) => SeekFrom::End(n),
+            std::io::SeekFrom::Current(n) => SeekFrom::Current(n),
+        }
+    }
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
@@ -162,6 +183,14 @@ impl Error for core::convert::Infallible {
 impl Error for ErrorKind {
     fn kind(&self) -> ErrorKind {
         *self
+    }
+}
+
+#[cfg(feature = "std")]
+#[cfg_attr(docsrs, doc(cfg(feature = "std")))]
+impl crate::Error for std::io::Error {
+    fn kind(&self) -> crate::ErrorKind {
+        self.kind().into()
     }
 }
